add: Skip Files/Directories Feature
You can specify a list of files in the template manifest to skip adding
to the output path. In other words these files/directoris will not be
processed through the template and skipped altogether. An example of a
file you would want to skip could be a image logo that is only for the
template marketing, or the .circleci or .github actions directories used
in automating releases.