add: ReplaceWith Feature
Specify a file that you want to subtitute in place of another during
processing (template pressing). Say you have a .circleci or .github
actions folder that you use for the template; but you also want to
generate a differenct .circleci or .github actions files in the
template output. You can add a `replace` section in the template
manifest in the format `"replace": [ "replace/.circleci:.circleci" ]`
that replaces the ".circleci" folder with the contents in
"replace/.circleci" in the final output.